Deal To Contact
Using Hubspot Client V3

This Node.js code works as follows: It retrieves the "e_mail_do_solicitante" property of the deal registered in the workflow, then stores the value of this property in the "emailDomain" variable and then creates a search criteria where it searches for contacts that have the same value as the "emailDomain" variable in the "email" property. If it finds results, the code will associate the Contact with the Deal.